pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Jhon Valderrama <jhonvalderramaa7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "0.1.1"
langchain-community = "0.0.13"
pinecone-client = "3.1.0"
pydantic-settings = "^2.5.2"
openai = "^1.50.2"
tiktoken = "^0.7.0"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-multipart = "^0.0.11"
shortuuid = "^1.0.13"
boto3 = "^1.35.29"
langfuse = "^2.51.2"
pytest = "^8.3.3"
coverage = "^7.6.1"
langchain-core = "0.1.9"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
